String->LocalDate Int->Float

+++ src/com/apmath/loans/domain/models/ApplicationInterface.kt
@@ -11,7 +11,7 @@ interface ApplicationInterface {
     val status: Status
 
     val clientId: Int
-    val interest: Int
+    val interest: Float
     val maxAmount: Money
     val minAmount: Money
 }
